requirements.txt
flask==3.0.2
opencv-python==4.9.0.80
pytube==15.0.0
openai-whisper==20231117
ollama==0.1.14
numpy==1.26.4